# Optimize formatting a bit